third_party/blink/web_tests/virtual/playback_speed_button/media/controls/overflow-button-disabled-no-source.html

<!DOCTYPE html>
<html>
<title>Test that the overflow menu is disabled with no source.</title>
<script src="../../../../resources/testharness.js"></script>
<script src="../../../../resources/testharnessreport.js"></script>
<script src="../../../../media/media-controls.js"></script>
<video controls width=400 controlsList=nodownload disablePictureInPicture></video>
<script>
async_test(t => {
  const video = document.querySelector('video');

  // Make sure the button is disabled.
  assert_true(overflowButton(video).disabled);
  assert_equals(overflowButton(video).style.display, '');

  // Set the source and start playing.
  video.src = '../../../../media/content/60_sec_video.webm';

  video.play().then(t.step_func_done(() => {
    // Make sure the button is no longer disabled.
    assert_false(overflowButton(video).disabled);
    assert_equals(overflowButton(video).style.display, '');
  }), t.unreached_func());
});
</script>
</html>
